cleanup after merge
